[0.23.0] Updates and backport process improvements (#1325)

+++ serving/docker/scripts/install_djl_serving.sh
@@ -16,9 +16,12 @@ if [ -z "$PYTORCH_JNI" ]; then
     unzip
 
   # install DJLServing
-  curl https://publish.djl.ai/djl-serving/djl-serving_${DJL_VERSION}-1_all.deb -f -o djl-serving_all.deb
+  if [ ! -f djl-serving_all.deb ]; then
+    curl https://publish.djl.ai/djl-serving/djl-serving_${DJL_VERSION}-1_all.deb -f -o djl-serving_all.deb
+  fi
   dpkg -i djl-serving_all.deb
   rm djl-serving_all.deb
+
   cp /usr/local/djl-serving-*/conf/log4j2.xml /opt/djl/conf/
   cp -r /usr/local/djl-serving-*/plugins /opt/djl/plugins
   # path api-0.23.0.jar for streaming timeout bug
